+++ crates/rpc/rpc-eth-api/src/helpers/call.rs
@@ -16,22 +16,21 @@ use reth_rpc_eth_types::{
     cache::db::{StateCacheDbRefMutWrapper, StateProviderTraitObjWrapper},
     error::ensure_success,
     revm_utils::{
-        apply_state_overrides, caller_gas_allowance, cap_tx_gas_limit_with_caller_allowance,
-        get_precompiles,
+        apply_block_overrides, apply_state_overrides, caller_gas_allowance,
+        cap_tx_gas_limit_with_caller_allowance, get_precompiles, CallFees,
     },
     EthApiError, EthResult, RevertError, RpcInvalidTransactionError, StateCacheDb,
 };
 use reth_rpc_server_types::constants::gas_oracle::{ESTIMATE_GAS_ERROR_RATIO, MIN_TRANSACTION_GAS};
 use reth_rpc_types::{
     state::{EvmOverrides, StateOverride},
-    AccessListWithGasUsed, BlockId, BlockOverrides, Bundle, EthCallResponse, StateContext,
-    TransactionInfo, TransactionRequest,
+    AccessListWithGasUsed, BlockId, Bundle, EthCallResponse, StateContext, TransactionInfo,
+    TransactionRequest,
 };
 use revm::{Database, DatabaseCommit};
 use revm_inspectors::access_list::AccessListInspector;
 #[cfg(feature = "optimism")]
 use revm_primitives::OptimismFields;
-use std::cmp::min;
 use tracing::trace;
 
 use super::{LoadBlock, LoadPendingBlock, LoadState, LoadTransaction, SpawnBlocking, Trace};
@@ -932,277 +931,5 @@ pub trait Call: LoadState + SpawnBlocking {
         }
 
         Ok(env)
-    }
-}
-
-/// Helper type for representing the fees of a [`TransactionRequest`]
-#[derive(Debug)]
-pub struct CallFees {
-    /// EIP-1559 priority fee
-    max_priority_fee_per_gas: Option<U256>,
-    /// Unified gas price setting
-    ///
-    /// Will be the configured `basefee` if unset in the request
-    ///
-    /// `gasPrice` for legacy,
-    /// `maxFeePerGas` for EIP-1559
-    gas_price: U256,
-    /// Max Fee per Blob gas for EIP-4844 transactions
-    max_fee_per_blob_gas: Option<U256>,
-}
-
-// === impl CallFees ===
-
-impl CallFees {
-    /// Ensures the fields of a [`TransactionRequest`] are not conflicting.
-    ///
-    /// # EIP-4844 transactions
-    ///
-    /// Blob transactions have an additional fee parameter `maxFeePerBlobGas`.
-    /// If the `maxFeePerBlobGas` or `blobVersionedHashes` are set we treat it as an EIP-4844
-    /// transaction.
-    ///
-    /// Note: Due to the `Default` impl of [`BlockEnv`] (Some(0)) this assumes the `block_blob_fee`
-    /// is always `Some`
-    fn ensure_fees(
-        call_gas_price: Option<U256>,
-        call_max_fee: Option<U256>,
-        call_priority_fee: Option<U256>,
-        block_base_fee: U256,
-        blob_versioned_hashes: Option<&[B256]>,
-        max_fee_per_blob_gas: Option<U256>,
-        block_blob_fee: Option<U256>,
-    ) -> EthResult<Self> {
-        /// Get the effective gas price of a transaction as specfified in EIP-1559 with relevant
-        /// checks.
-        fn get_effective_gas_price(
-            max_fee_per_gas: Option<U256>,
-            max_priority_fee_per_gas: Option<U256>,
-            block_base_fee: U256,
-        ) -> EthResult<U256> {
-            match max_fee_per_gas {
-                Some(max_fee) => {
-                    if max_fee < block_base_fee {
-                        // `base_fee_per_gas` is greater than the `max_fee_per_gas`
-                        return Err(RpcInvalidTransactionError::FeeCapTooLow.into())
-                    }
-                    if max_fee < max_priority_fee_per_gas.unwrap_or(U256::ZERO) {
-                        return Err(
-                            // `max_priority_fee_per_gas` is greater than the `max_fee_per_gas`
-                            RpcInvalidTransactionError::TipAboveFeeCap.into(),
-                        )
-                    }
-                    Ok(min(
-                        max_fee,
-                        block_base_fee
-                            .checked_add(max_priority_fee_per_gas.unwrap_or(U256::ZERO))
-                            .ok_or_else(|| {
-                                EthApiError::from(RpcInvalidTransactionError::TipVeryHigh)
-                            })?,
-                    ))
-                }
-                None => Ok(block_base_fee
-                    .checked_add(max_priority_fee_per_gas.unwrap_or(U256::ZERO))
-                    .ok_or_else(|| EthApiError::from(RpcInvalidTransactionError::TipVeryHigh))?),
-            }
-        }
-
-        let has_blob_hashes =
-            blob_versioned_hashes.as_ref().map(|blobs| !blobs.is_empty()).unwrap_or(false);
-
-        match (call_gas_price, call_max_fee, call_priority_fee, max_fee_per_blob_gas) {
-            (gas_price, None, None, None) => {
-                // either legacy transaction or no fee fields are specified
-                // when no fields are specified, set gas price to zero
-                let gas_price = gas_price.unwrap_or(U256::ZERO);
-                Ok(Self {
-                    gas_price,
-                    max_priority_fee_per_gas: None,
-                    max_fee_per_blob_gas: has_blob_hashes.then_some(block_blob_fee).flatten(),
-                })
-            }
-            (None, max_fee_per_gas, max_priority_fee_per_gas, None) => {
-                // request for eip-1559 transaction
-                let effective_gas_price = get_effective_gas_price(
-                    max_fee_per_gas,
-                    max_priority_fee_per_gas,
-                    block_base_fee,
-                )?;
-                let max_fee_per_blob_gas = has_blob_hashes.then_some(block_blob_fee).flatten();
-
-                Ok(Self {
-                    gas_price: effective_gas_price,
-                    max_priority_fee_per_gas,
-                    max_fee_per_blob_gas,
-                })
-            }
-            (None, max_fee_per_gas, max_priority_fee_per_gas, Some(max_fee_per_blob_gas)) => {
-                // request for eip-4844 transaction
-                let effective_gas_price = get_effective_gas_price(
-                    max_fee_per_gas,
-                    max_priority_fee_per_gas,
-                    block_base_fee,
-                )?;
-                // Ensure blob_hashes are present
-                if !has_blob_hashes {
-                    // Blob transaction but no blob hashes
-                    return Err(RpcInvalidTransactionError::BlobTransactionMissingBlobHashes.into())
-                }
-
-                Ok(Self {
-                    gas_price: effective_gas_price,
-                    max_priority_fee_per_gas,
-                    max_fee_per_blob_gas: Some(max_fee_per_blob_gas),
-                })
-            }
-            _ => {
-                // this fallback covers incompatible combinations of fields
-                Err(EthApiError::ConflictingFeeFieldsInRequest)
-            }
-        }
-    }
-}
-
-/// Applies the given block overrides to the env
-fn apply_block_overrides(overrides: BlockOverrides, env: &mut BlockEnv) {
-    let BlockOverrides {
-        number,
-        difficulty,
-        time,
-        gas_limit,
-        coinbase,
-        random,
-        base_fee,
-        block_hash: _,
-    } = overrides;
-
-    if let Some(number) = number {
-        env.number = number;
-    }
-    if let Some(difficulty) = difficulty {
-        env.difficulty = difficulty;
-    }
-    if let Some(time) = time {
-        env.timestamp = U256::from(time);
-    }
-    if let Some(gas_limit) = gas_limit {
-        env.gas_limit = U256::from(gas_limit);
-    }
-    if let Some(coinbase) = coinbase {
-        env.coinbase = coinbase;
-    }
-    if let Some(random) = random {
-        env.prevrandao = Some(random);
-    }
-    if let Some(base_fee) = base_fee {
-        env.basefee = base_fee;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use reth_primitives::constants::GWEI_TO_WEI;
-
-    #[test]
-    fn test_ensure_0_fallback() {
-        let CallFees { gas_price, .. } =
-            CallFees::ensure_fees(None, None, None, U256::from(99), None, None, Some(U256::ZERO))
-                .unwrap();
-        assert!(gas_price.is_zero());
-    }
-
-    #[test]
-    fn test_blob_fees() {
-        let CallFees { gas_price, max_fee_per_blob_gas, .. } =
-            CallFees::ensure_fees(None, None, None, U256::from(99), None, None, Some(U256::ZERO))
-                .unwrap();
-        assert!(gas_price.is_zero());
-        assert_eq!(max_fee_per_blob_gas, None);
-
-        let CallFees { gas_price, max_fee_per_blob_gas, .. } = CallFees::ensure_fees(
-            None,
-            None,
-            None,
-            U256::from(99),
-            Some(&[B256::from(U256::ZERO)]),
-            None,
-            Some(U256::from(99)),
-        )
-        .unwrap();
-        assert!(gas_price.is_zero());
-        assert_eq!(max_fee_per_blob_gas, Some(U256::from(99)));
-    }
-
-    #[test]
-    fn test_eip_1559_fees() {
-        let CallFees { gas_price, .. } = CallFees::ensure_fees(
-            None,
-            Some(U256::from(25 * GWEI_TO_WEI)),
-            Some(U256::from(15 * GWEI_TO_WEI)),
-            U256::from(15 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        )
-        .unwrap();
-        assert_eq!(gas_price, U256::from(25 * GWEI_TO_WEI));
-
-        let CallFees { gas_price, .. } = CallFees::ensure_fees(
-            None,
-            Some(U256::from(25 * GWEI_TO_WEI)),
-            Some(U256::from(5 * GWEI_TO_WEI)),
-            U256::from(15 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        )
-        .unwrap();
-        assert_eq!(gas_price, U256::from(20 * GWEI_TO_WEI));
-
-        let CallFees { gas_price, .. } = CallFees::ensure_fees(
-            None,
-            Some(U256::from(30 * GWEI_TO_WEI)),
-            Some(U256::from(30 * GWEI_TO_WEI)),
-            U256::from(15 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        )
-        .unwrap();
-        assert_eq!(gas_price, U256::from(30 * GWEI_TO_WEI));
-
-        let call_fees = CallFees::ensure_fees(
-            None,
-            Some(U256::from(30 * GWEI_TO_WEI)),
-            Some(U256::from(31 * GWEI_TO_WEI)),
-            U256::from(15 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        );
-        assert!(call_fees.is_err());
-
-        let call_fees = CallFees::ensure_fees(
-            None,
-            Some(U256::from(5 * GWEI_TO_WEI)),
-            Some(U256::from(GWEI_TO_WEI)),
-            U256::from(15 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        );
-        assert!(call_fees.is_err());
-
-        let call_fees = CallFees::ensure_fees(
-            None,
-            Some(U256::MAX),
-            Some(U256::MAX),
-            U256::from(5 * GWEI_TO_WEI),
-            None,
-            None,
-            Some(U256::ZERO),
-        );
-        assert!(call_fees.is_err());
     }
 }
